Changing logging from StdLog to log4perl
StdLog doesn't seem to be maintained, and while it's available
in CPAN, it's not shipped with anything normally.  This is a quick
sledge hammer fix to switch to log4perl, which is reasonably
widely used and still maintained.

This also removes the test_log.pl script, since it duplicates
functionality, and frankly, isn't useful